Reformat params, knobs, and sliders. Star skeleton
Sliders and knob functionality is transferrable to other scenes, and
corresponding actions can be individually defined.

Params are now stored in separate objects similar to synths.js for ease
of use and to not have to redefine each time a new scene is called only
to immediately write the saved parameter values to them.

All star buttons have been added but are still unusable.